WIP add more checks

+++ src/xrpld/rpc/handlers/Tx.cpp
@@ -153,7 +153,7 @@ doTxHelp(RPC::Context& context, TxArgs args)
             uint32_t netID = context.app.getNetworkIDService().getNetworkID();
 
             if (txnIdx <= 0xFFFFU && netID < 0xFFFFU && lgrSeq < 0x0FFF'FFFFUL)
-                result.ctid = RPC::encodeCTID(lgrSeq, (uint32_t)txnIdx, (uint32_t)netID);
+                result.ctid = RPC::encodeCTID(lgrSeq, txnIdx, netID);
         }
     }
